*prevent deletion of default `namespace` and default `experiment`.

+++ pkg/api/admin/service/namespace/service.go
@@ -99,6 +99,9 @@ func (s Service) DeleteNamespace(ctx context.Context, id uint) error {
 	if namespace == nil {
 		return eris.Errorf("namespace not found by id: %d", id)
 	}
+	if namespace.IsDefault() {
+		return eris.Errorf("unable to delete default namespace")
+	}
 	if err := s.namespaceRepository.Delete(ctx, namespace); err != nil {
 		return eris.Wrap(err, "error deleting namespace")
 	}
